add collection removal

+++ tests/test_write.py
@@ -209,6 +209,7 @@ def test_multiprocess_insert(test_client, test_collection_name):
     NUM_OF_DOCUMENTS_INSERTED = 10
     if test_collection_name in test_client.list_collections():    
         test_client.delete_collection(test_collection_name)
+        time.sleep(5)
     documents = test_client.create_sample_documents(NUM_OF_DOCUMENTS_INSERTED)
     results = test_client.insert_documents(test_collection_name, documents, workers=5)
     time.sleep(10)
@@ -241,6 +242,7 @@ def test_multiprocess_insert_with_error_with_overwrite(test_client, test_collect
     NUM_OF_DOCUMENTS_INSERTED = 100
     if test_collection_name in test_client.list_collections():    
         test_client.delete_collection(test_collection_name)
+        time.sleep(5)
     documents = test_client.create_sample_documents(NUM_OF_DOCUMENTS_INSERTED)
     documents.append({
         '_id': '9993',
@@ -260,6 +262,7 @@ def test_multiprocess_with_collection_client(test_collection_client, test_collec
     NUM_OF_DOCUMENTS_INSERTED = 10
     if test_collection_client.collection_name in test_collection_client.list_collections():    
         test_collection_client.delete_collection()
+        time.sleep(5)
     documents = test_collection_client.create_sample_documents(NUM_OF_DOCUMENTS_INSERTED)
     results = test_collection_client.insert_documents(documents, workers=5)
     time.sleep(10)
@@ -273,6 +276,7 @@ def test_multiprocess__with_error_with_collection_client(test_collection_client)
     NUM_OF_DOCUMENTS_INSERTED = 10
     if test_collection_client.collection_name in test_collection_client.list_collections():    
         test_collection_client.delete_collection()
+        time.sleep(5)
     documents = test_collection_client.create_sample_documents(NUM_OF_DOCUMENTS_INSERTED)
     documents.append({
         '_id': 3,
@@ -283,11 +287,13 @@ def test_multiprocess__with_error_with_collection_client(test_collection_client)
     time.sleep(10)
     assert len(results['failed_document_ids']) == 1
     assert test_collection_client.collection_name in test_collection_client.list_collections()
-    assert test_collection_client.collection_stats()['number_of_documents'] == NUM_OF_DOCUMENTS_INSERTED
-    test_collection_client.delete_collection()
+    assert test_collection_client.collection_stats()['number_of_documents'] == NUM_OF_DOCUMENTS_INSERTED 
 
 @pytest.mark.use_client
 def test_multiprocess_with_overwrite(test_client, test_collection_name):
+    if test_collection_client.collection_name in test_collection_client.list_collections():    
+        test_collection_client.delete_collection()
+        time.sleep(5)
     NUM_OF_DOCS = 10
     docs = test_client.create_sample_documents(NUM_OF_DOCS)
     test_client.insert_documents(test_collection_name, docs[0:5], workers=1, overwrite=False)
@@ -297,6 +303,9 @@ def test_multiprocess_with_overwrite(test_client, test_collection_name):
 
 @pytest.mark.use_client
 def test_multiprocess_with_overwrite_insert(test_client, test_collection_name):
+    if test_collection_client.collection_name in test_collection_client.list_collections():    
+        test_collection_client.delete_collection()
+        time.sleep(5)
     NUM_OF_DOCS = 10
     docs = test_client.create_sample_documents(NUM_OF_DOCS)
     test_client.insert_documents(test_collection_name, docs[0:5], workers=1, overwrite=False)
@@ -306,6 +315,9 @@ def test_multiprocess_with_overwrite_insert(test_client, test_collection_name):
 
 @pytest.mark.use_client
 def test_multiprocess_overwrite(test_client, test_collection_name):
+    if test_collection_client.collection_name in test_collection_client.list_collections():    
+        test_collection_client.delete_collection()
+        time.sleep(5)
     NUM_OF_DOCS = 10
     docs = test_client.create_sample_documents(NUM_OF_DOCS)
     test_client.insert_documents(test_collection_name, docs[0:5], workers=1, overwrite=False)
@@ -325,6 +337,9 @@ def test_multiprocess_overwrite(test_client, test_collection_name):
 
 @pytest.mark.use_client
 def test_multiprocess_not_overwrite(test_client, test_collection_name):
+    if test_collection_client.collection_name in test_collection_client.list_collections():    
+        test_collection_client.delete_collection()
+        time.sleep(5)
     NUM_OF_DOCS = 10
     docs = test_client.create_sample_documents(NUM_OF_DOCS)
     test_client.insert_documents(test_collection_name, docs[0:5], workers=1, overwrite=False)
@@ -344,6 +359,9 @@ def test_multiprocess_not_overwrite(test_client, test_collection_name):
 
 @pytest.mark.use_client
 def test_multiprocess_overwrite_collection_client(test_collection_client, test_collection_name):
+    if test_collection_client.collection_name in test_collection_client.list_collections():    
+        test_collection_client.delete_collection()
+        time.sleep(5)
     NUM_OF_DOCS = 10
     docs = test_collection_client.create_sample_documents(NUM_OF_DOCS)
     test_collection_client.insert_documents(docs[0:5], workers=1, overwrite=False)
